+++ src/ploomber/env/env.py
@@ -1,17 +1,16 @@
 """
 Environment management
 """
-import weakref
+import pydoc
 import logging
-from warnings import warn
 from itertools import chain
 from pathlib import Path
 from glob import iglob
 from io import StringIO
 import getpass
-import tempfile
 import platform
 from functools import partial
+from collections.abc import Mapping
 
 from ploomber.FrozenJSON import FrozenJSON
 from ploomber.path import PathManager
@@ -21,7 +20,7 @@ import yaml
 from jinja2 import Template
 
 
-# TODO: add defaults arg
+# TODO: add defaults functionality if defined in env.yaml
 # TODO: improve str(Env) and repr(Env)
 
 
@@ -48,8 +47,9 @@ class Env:
     interpreted as directories and they will be created if they do not exist
 
     There are two wildcards available "{{user}}" (returns the current user)
-    and "{{git_location}}" (if the env.yaml file is located inside a git
-    repo, this will return the current branch name, if in detached HEAD
+    and "{{git_location}}" (if the env.yaml file has a "module" key, then
+    # TODO: edit this
+    ... this will return the current branch name, if in detached HEAD
     state, it will return the hash to the current commit
 
     Notes
@@ -60,7 +60,8 @@ class Env:
     in Env objects are meant to be constants (such as db URIs) there should not
     be a need for multiple Envs to exist at any given time. While it is
     possible to switch to a different Env in the same process, this is
-    discouraged.
+    discouraged since that could lead to subtle bugs if modules set variables
+    from Env parameters
 
 
     Examples
@@ -69,69 +70,65 @@ class Env:
     Basic usage
 
     >>> from ploomber import Env
+    >>> Env.start({'a': 1, 'module': 'my_project'})
     >>> env = Env()
     >>> env.db.uri # traverse the yaml tree structure using dot notation
     >>> env.path.raw # returns an absolute path to the raw data
     """
-    _instances = set()
+    _data = None
 
-    __path_to_env = None
-    __wildcards_replace = {}
-
-    def __init__(self, path_to_env=None):
-        self._logger = logging.getLogger(__name__)
-
-        # if not env has been set...
-        if Env.__path_to_env is None:
+    @classmethod
+    def start(cls, source=None):
+        if cls._data is None:
 
             # try to set it if no argument was provided
-            if path_to_env is None:
-
+            if source is None:
                 # look for an env.{name}.yaml, if that fails, try env.yaml
                 name = platform.node()
-                path_to_env = find_env_w_name(name)
+                path_found = find_env_w_name(name)
 
-                if path_to_env is None:
+                if path_found is None:
+                    # TODO: improve error message, include hostname
                     raise FileNotFoundError("Couldn't find env.yaml")
+                else:
+                    source = path_found
 
-            # resolve it
-            path_to_env = Path(path_to_env).resolve()
+            if isinstance(source, (str, Path)):
+                cls._path_to_env = Path(source).resolve()
+                cls._name = _get_name(cls._path_to_env)
+            else:
+                cls._path_to_env = None
+                cls._name = None
 
-            # and save a reference to the file
-            Env.__path_to_env = path_to_env
+            cls._path = PathManager(cls)
+            cls._data = load(source)
+
+            return cls()
 
         # if an environment has been set...
         else:
-            # if no argument provided, just use whatever env was already used
-            if path_to_env is None:
-                path_to_env = Env.__path_to_env
-                self._logger.info('Already instantiated Env, loading it from '
-                                  f'{path_to_env}...')
+            raise RuntimeError('Cannot start environment, one has already '
+                               'started: {}'.format(cls()))
 
-            # otherwise, resolve argument and warn the user if there is
-            # conflcit
-            else:
-                path_to_env = Path(path_to_env).resolve()
+    @classmethod
+    def end(cls):
+        cls._path_to_env = None
+        cls._name = None
+        cls._path = None
+        cls._data = None
 
-                if Env.__path_to_env != path_to_env:
-                    warn('Env was already instantiated using file '
-                         f'{Env.__path_to_env} but new '
-                         'instance was created using '
-                         f'{path_to_env}, it is not recommended to '
-                         'have more than one environment per project')
+    def __init__(self):
+        if self._data is None:
+            raise RuntimeError('Env has not been set, run Env.start before '
+                               'running Env()')
 
-        self._path_to_env = path_to_env
-        self._name = _get_name(path_to_env)
-        self._path = PathManager(path_to_env, self)
-
-        self._env_content = self.load(path_to_env)
-        self._instances.add(weakref.ref(self))
+        self._logger = logging.getLogger(__name__)
 
     def __repr__(self):
-        return f'Env loaded from {self._path_to_env}'
+        return f'Env: {self._path_to_env} - {repr(self._data)}'
 
     def __dir__(self):
-        return dir(self._env_content)
+        return dir(self._data)
 
     @property
     def name(self):
@@ -140,15 +137,15 @@ class Env:
     @property
     def path(self):
         """
-        path manager, returnd paths to directories specified in your env
+        path manager, return paths to directories specified in your env
         """
         return self._path
 
     def __getattr__(self, key):
-        return getattr(self._env_content, key)
+        return getattr(self._data, key)
 
     def __getitem__(self, key):
-        return self._env_content[key]
+        return self._data[key]
 
     def __setattr__(self, name, value):
         if name.startswith('_'):
@@ -161,63 +158,68 @@ class Env:
         """
         return repo.get_env_metadata(self.path.home)
 
-    def load(self, path_to_env):
-        path_to_env = Path(path_to_env)
-        home = path_to_env.parent
-        env_content = path_to_env.read_text()
 
-        params = dict(user=self.__wildcards_replace.get(
-            'user') or getpass.getuser())
+def load(source):
+    if isinstance(source, (str, Path)):
+        env_content = Path(source).read_text()
 
-        # only try to find git location if {{git_location is used}}
-        if '{{git_location}}' in env_content:
-            if self.__wildcards_replace.get('git_location'):
-                params['git_location'] = self.__wildcards_replace['git_location']
+        with StringIO(env_content) as f:
+            m = yaml.load(f, Loader=yaml.SafeLoader)
+
+        module_name = m.get('module')
+    elif isinstance(source, Mapping):
+        env_content = yaml.dump(source)
+        module_name = source.get('module')
+
+    # if "module" is defined in the env file, make sure you can import it
+    if module_name is None:
+        module = None
+    else:
+        module = pydoc.locate(module_name)
+
+        if module is None:
+            raise ImportError('Could not import module "{}"'
+                              .format(module_name))
+
+    # at this point if "module" was defined, we have the module object,
+    # otherwise both are None
+
+    params = dict(user=getpass.getuser())
+
+    if '{{version}}' in env_content:
+        if module_name is None:
+            raise RuntimeError('The git_location placeholder is only available '
+                               'if Env defines a "module" constant')
+        else:
+            if hasattr(module, '__version__'):
+                params['version'] = module.__version__
             else:
-                params['git_location'] = (repo
-                                          .get_env_metadata(home)['git_location'])
+                raise RuntimeError('Module {} does not have a __version__, cannot '
+                             'expand version placeholder'.format(module_name))
 
-        s = Template(env_content).render(**params)
+    if '{{git_location}}' in env_content:
+        if module_name is None:
+            raise RuntimeError('The git_location placeholder is only available '
+                               'if Env defines a "module" constant')
+        else:
+            module_path = str(Path(module.__file__).parent.absolute())
+            params['git_location'] = (repo
+                                      .get_env_metadata(module_path)
+                                      ['git_location'])
 
-        with StringIO(s) as f:
-            content = yaml.load(f, Loader=yaml.SafeLoader)
+    s = Template(env_content).render(**params)
 
-        env = FrozenJSON(content)
+    with StringIO(s) as f:
+        content = yaml.load(f, Loader=yaml.SafeLoader)
 
-        return env
+    env = FrozenJSON(content)
 
-    @classmethod
-    def getinstances(cls):
-        # TODO: use a singleton instead of this?
-        # http://effbot.org/pyfaq/how-do-i-get-a-list-of-all-instances-of-a-given-class.htm
-        dead = set()
-        for ref in cls._instances:
-            obj = ref()
-            if obj is not None:
-                yield obj
-            else:
-                dead.add(ref)
-        cls._instances -= dead
-
-    @classmethod
-    def _set_wildcards(cls, d):
-        cls.__wildcards_replace = d
-        # re-initialize content
-        for obj in cls.getinstances():
-            obj._env_content = obj.load(obj._path_to_env)
-
-    @classmethod
-    def from_dict(cls, d):
-        # FIXME: this method should not rely on yaml files
-        _, file = tempfile.mkstemp(prefix='env.', suffix='.yaml')
-
-        with open(file, 'w') as f:
-            yaml.dump(d, f)
-
-        return cls(file)
+    return env
 
 
 def find_env_w_name(name):
+    """Find environment named env.{name}.yaml by going to parent folders
+    """
     path = find_env(name='env.{}.yaml'.format(name))
 
     if path is None:
@@ -227,6 +229,8 @@ def find_env_w_name(name):
 
 
 def find_env(name, max_levels_up=6):
+    """Find environment by going to the parent folders
+    """
     def levels_up(n):
         return chain.from_iterable(iglob('../' * i + '**')
                                    for i in range(n + 1))
@@ -243,19 +247,9 @@ def find_env(name, max_levels_up=6):
     return path_to_env
 
 
-def load_config(config_file):
-    """Loads configuration file, asumes file is in config/
-
-    Parameters
-    ----------
-    file: str
-        As returned from __file__
-    """
-    project_dir = Env().project_dir
-    return Path(project_dir, 'config', config_file).absolute()
-
-
 def _get_name(path_to_env):
+    """Parse env.{name}.yaml -> name
+    """
     filename = str(Path(path_to_env).name)
 
     err = ValueError('Wrong filename, must be either env.{name}.yaml '