+++ src/ploomber/path.py
@@ -3,18 +3,11 @@ from pathlib import Path
 
 class PathManager:
 
-    def __init__(self, path_to_env, env):
-        self._home = Path(path_to_env).resolve().parent
+    def __init__(self, env):
         self._env = env
 
-    @property
-    def home(self):
-        """Project's home folder
-        """
-        return self._home
-
     def __getattr__(self, key):
-        raw_value = getattr(self._env._env_content.path, key)
+        raw_value = getattr(self._env._data.path, key)
         path = Path(raw_value)
 
         # need this if statement in case the path is a file, otherwise it will